Move AddonsState class to a dedicated module
